backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1